+++ examples/led_actor/button_actor.h
@@ -5,9 +5,9 @@
 #pragma once
 #include <ipc.h>
 
-IPC_EVENT_DEFINE(ButtonClick, { uint8_t button_id; });
-IPC_EVENT_DEFINE(ButtonDoubleClick, { uint8_t button_id; });
-IPC_EVENT_DEFINE(ButtonHold, {
+IPC_EVENT_DEFINE_LOCAL(ButtonClick, { uint8_t button_id; });
+IPC_EVENT_DEFINE_LOCAL(ButtonDoubleClick, { uint8_t button_id; });
+IPC_EVENT_DEFINE_LOCAL(ButtonHold, {
     uint8_t button_id;
     uint32_t hold_ms;
 });